requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6
sib-api-v3-sdk>=7.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
slowapi>=0.1.9
